(split)Another series of warnings for aggressive compilers
git-svn-id: http://svn.xvid.org/trunk@723 060a26f8-e6a2-0410-8959-474133d486e0

+++ src/quant/quant_mpeg4.c
@@ -50,7 +50,7 @@
  *  exception also makes it possible to release a modified version which
  *  carries forward this exception.
  *
- * $Id: quant_mpeg4.c,v 1.7 2002-11-26 23:44:11 edgomez Exp $
+ * $Id: quant_mpeg4.c,v 1.8 2002-12-15 01:21:12 edgomez Exp $
  *
  ****************************************************************************/
 
@@ -122,7 +122,7 @@ quant4_intra_c(int16_t * coeff,
 
 	intra_matrix = get_intra_matrix();
 
-	coeff[0] = DIV_DIV(data[0], (int32_t) dcscalar);
+	coeff[0] = (int16_t)(DIV_DIV(data[0], (int32_t) dcscalar));
 
 	for (i = 1; i < 64; i++) {
 		if (data[i] < 0) {
@@ -136,7 +136,7 @@ quant4_intra_c(int16_t * coeff,
 
 			level = ((level << 4) + (intra_matrix[i] >> 1)) / intra_matrix[i];
 			level = ((level + quantd) * mult) >> 17;
-			coeff[i] = level;
+			coeff[i] = (int16_t)level;
 		} else {
 			coeff[i] = 0;
 		}
@@ -160,7 +160,7 @@ dequant4_intra_c(int16_t * data,
 
 	intra_matrix = get_intra_matrix();
 
-	data[0] = coeff[0] * dcscalar;
+	data[0] = (int16_t)(coeff[0] * dcscalar);
 	if (data[0] < -2048) {
 		data[0] = -2048;
 	} else if (data[0] > 2047) {
@@ -180,7 +180,7 @@ dequant4_intra_c(int16_t * data,
 			uint32_t level = coeff[i];
 
 			level = (level * intra_matrix[i] * quant) >> 3;
-			data[i] = (level <= 2047 ? level : 2047);
+			data[i] = (int16_t)(level <= 2047 ? level : 2047);
 		}
 	}
 }
@@ -221,7 +221,7 @@ quant4_inter_c(int16_t * coeff,
 			level = ((level << 4) + (inter_matrix[i] >> 1)) / inter_matrix[i];
 			level = (level * mult) >> 17;
 			sum += level;
-			coeff[i] = level;
+			coeff[i] = (int16_t)level;
 		} else {
 			coeff[i] = 0;
 		}
@@ -259,7 +259,7 @@ dequant4_inter_c(int16_t * data,
 			uint32_t level = coeff[i];
 
 			level = ((2 * level + 1) * inter_matrix[i] * quant) >> 4;
-			data[i] = (level <= 2047 ? level : 2047);
+			data[i] = (int16_t)(level <= 2047 ? level : 2047);
 		}
 
 		sum ^= data[i];
